requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
async-generator==1.10
attrs==22.1.0
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
h11==0.14.0
html-testRunner==1.2.1
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-html==3.1.1
pytest-metadata==2.0.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
self==2020.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
